Move HiveCatalogFileIndex back into CatalogFileIndex
